# refactor(query): Optimize the `array_agg` function to reduce memory usage