--- a/src/query/expression/src/values.rs
+++ b/src/query/expression/src/values.rs
@@ -514,6 +514,18 @@ impl Scalar {
             _ => None,
         }
     }
+
+    pub fn get_slice(&self) -> Option<&[u8]> {
+        match self {
+            Scalar::String(val) => Some(val.as_bytes()),
+            Scalar::Geography(val) => Some(val.0.as_slice()),
+            Scalar::Binary(val)
+            | Scalar::Bitmap(val)
+            | Scalar::Variant(val)
+            | Scalar::Geometry(val) => Some(val.as_slice()),
+            _ => None,
+        }
+    }
 }
 
 impl ScalarRef<'_> {