… src/query/functions/src/aggregates/aggregate_array_agg.rs
… src/query/functions/tests/it/aggregates/testdata/agg.txt
… src/query/functions/tests/it/aggregates/testdata/agg_group_by.txt