I hereby agree to the terms of the CLA available at: https://docs.databend.com/dev/policies/cla/

## Summary

**Description:**

This PR optimizes the `array_agg` function to significantly reduce memory consumption and improve performance, particularly when aggregating large datasets.

**Key Changes:**

*   **Replaced `Vec<T::Scalar>` and `Vec<Option<T::Scalar>>` with `BinaryColumn`:** The internal implementation of `array_agg` now uses `BinaryColumn` to store aggregated values. This change drastically reduces memory footprint, preventing potential Out-of-Memory (OOM) errors.
*   **NULL Value Handling:** The implementation now efficiently ignores NULL values within the array during aggregation, avoiding unnecessary processing and storage.

**Performance Impact:**

The following example demonstrates the performance improvements using the provided test case:

```sql
CREATE OR REPLACE TABLE user_activity_logs AS
SELECT 
    number % 100000000 AS id,
    JSON_OBJECT(
        'id', CAST(number % 100000000 AS STRING),
        'email', CONCAT('user', CAST(number % 100000000 AS STRING), '@example.com')
    ) AS data
FROM numbers(5000000);

SELECT array_agg(data:id) FROM user_activity_logs IGNORE_RESULT;
SELECT array_agg(data:email) FROM user_activity_logs IGNORE_RESULT;
```


| Query                 | Memory Usage (Before) | Memory Usage (After) | Execution Time (Before) | Execution Time (After) |
| :-------------------- | :------: | -------: | :------: | -------: |
| array_agg(data:id)    | 262MB  | 108MB  | 4.325 sec | 1.017 sec |
| array_agg(data:email) | 338MB  | 184MB  | 4.427 sec | 1.112 sec |


- fixes: #[Link the issue here]

## Tests

- [x] Unit Test
- [x] Logic Test
- [ ] Benchmark Test
- [ ] No Test - _Explain why_

## Type of change

- [ ] Bug Fix (non-breaking change which fixes an issue)
- [ ] New Feature (non-breaking change which adds functionality)
- [ ] Breaking Change (fix or feature that could cause existing functionality not to work as expected)
- [ ] Documentation Update
- [x] Refactoring
- [ ] Performance Improvement
- [ ] Other (please describe):